feat: refactor component name, add custom date/time hook and account/help dropdown buttons
